common/org.cheetahplatform.test/src/org/cheetahplatform/test/modeler/RenameCommandTest.java
/*******************************************************************************
 * This file is provided under the terms of the Eclipse Public License 1.0. 
 * Any use, reproduction or distribution of the program
 * constitutes recipient's acceptance of this agreement.
 *******************************************************************************/

package org.cheetahplatform.test.modeler;

import static org.cheetahplatform.modeler.graph.command.AbstractGraphCommand.RENAME;
import static org.cheetahplatform.test.modeler.AbstractGraphCommandTest.NODE;
import static org.junit.Assert.assertEquals;

import java.util.ArrayList;

import org.cheetahplatform.modeler.graph.command.RenameCommand;
import org.cheetahplatform.modeler.graph.descriptor.IGraphElementDescriptor;
import org.cheetahplatform.modeler.graph.model.Graph;
import org.cheetahplatform.modeler.graph.model.Node;
import org.junit.Test;

public class RenameCommandTest {
	@Test
	public void undo() throws Exception {
		Graph graph = new Graph(new ArrayList<IGraphElementDescriptor>());
		TestLogListener listener = new TestLogListener();
		graph.addLogListener(listener);
		Node node = NODE.createModel(graph);
		RenameCommand command = new RenameCommand(node, "new name");
		command.execute();

		assertEquals("new name", node.getName());
		assertEquals(1, listener.getEntries().size());
		assertEquals(RENAME, listener.getEntries().get(0).getEventType());

		command.undo();
		assertEquals(null, node.getName());
		assertEquals(2, listener.getEntries().size());
		assertEquals(RENAME, listener.getEntries().get(1).getEventType());
	}
}
